refactor: move commands::kv::bucket to own mod sites
