+++ src/sites/manifest.rs
